backend/requirements.txt
Django==4.1.7
djangorestframework
django-cors-headers
psycopg2
